v0.4.0 Phase 45: Tech Debt Cleanup

+++ crates/assay-types/src/gate_run.rs
@@ -65,12 +65,12 @@ inventory::submit! {
 ///
 /// Present only when truncation occurred (diff exceeded token budget).
 /// Omitted entirely when the diff fit within budget (clean passthrough).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema)]
 pub struct DiffTruncation {
     /// Byte size of the original diff before truncation.
-    pub original_bytes: usize,
+    pub original_bytes: u64,
     /// Byte size of the diff after truncation.
-    pub truncated_bytes: usize,
+    pub truncated_bytes: u64,
     /// Files included in the truncated diff.
     pub included_files: Vec<String>,
     /// Files omitted from the truncated diff (present in original, absent after truncation).
